pyproject.toml
[project]
name = "ask-your-second-brain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.12.0",
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.6",
    "langchain-huggingface>=1.2.0",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.52.2",
    "transformers>=4.57.3",
]
